add second led test

+++ Buzzer/Core/Inc/led_pwm.h
@@ -12,7 +12,7 @@
 
 void enable_timer3(void);
 
-void generate_signal(uint8_t* data, uint32_t data_length);
+void generate_signal(uint16_t* data, uint32_t data_length);
 
 void dma_init(void);
