+++ Buzzer/Core/Src/led_pwm.c
@@ -14,7 +14,7 @@ void enable_timer3(){
 }
 
 
-void generate_signal(uint8_t* data, uint32_t data_length){
+void generate_signal(uint16_t* data, uint32_t data_length){
 	LL_DMA_ConfigAddresses(DMA1, LL_DMA_CHANNEL_6, (uint32_t)data, (uint32_t)&TIM3->CCR1, LL_DMA_DIRECTION_MEMORY_TO_PERIPH);
 	LL_DMA_SetDataLength(DMA1, LL_DMA_CHANNEL_6, data_length);
 	LL_DMA_EnableIT_TC(DMA1, LL_DMA_CHANNEL_6);
